% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank.R
\docType{data}
\name{bank}
\alias{bank}
\title{Bank Customer Data}
\format{
A tibble with 100 rows and 2 columns:
\describe{
\item{arrival_time}{Arrival time of the customers (in seconds).}
\item{service_time}{Service time for each customer (in seconds).}
}
}
\source{
Simulated data.
}
\usage{
bank
}
\description{
A dataset containing simulated arrival times and service times.
}
\keyword{datasets}
