% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Multiserver}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{The arrival time of the customers}

\item{ServiceTimes}{The service time of the customers}

\item{NumServers}{The number of customers served}
}
\value{
Returns the tibble of the data
}
\description{
A function which takes in the inputs and returns a tibble of the data
}
\examples{
data <- matrix(data = 1:20, nrow = 10, ncol = 2)
Multiserver(data[,1],data[,2])
}
